Add ListWise Integration
create a simple module that allows you to check if an email is clean by using
the ListWise api. You can also check its general status as a side bonus!.

I used the 'httparty' gem because I've had much better experience with it
in the past, compared to the 'net/http' library.